requirements.txt
accelerate==1.1.1
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1748348059416/work
acres==0.5.0
aiofiles==24.1.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==26.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==7.0.5
certifi==2025.6.15
cffi==2.0.0
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1741797914774/work
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.2
cycler==0.12.1
dcor==0.6
dill==0.4.0
distro==1.9.0
emd==0.8.1
EMD-signal==1.6.4
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
faiss-cpu==1.13.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py @ file:///D:/bld/h5py_1749298248358/work
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1747722781675/work
kiwisolver==1.4.8
langgraph==0.2.67
libclang==18.1.1
llvmlite==0.45.0
looseversion==1.3.0
lxml==6.0.2
Markdown==3.8.1
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
ml-dtypes @ file:///D:/bld/ml_dtypes_1736538822829/work
mpmath==1.3.0
multiprocess==0.70.18
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1748346368839/work
narwhals==2.19.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numba==0.62.0
numpy==2.2.6
openai==2.32.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree @ file:///D:/bld/optree_1748442476563/work
packaging==24.2
pandas==2.3.2
pathlib==1.0.1
pathos==0.3.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pox==0.3.6
ppft==1.7.7
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyarrow==23.0.1
pycparser==2.23
pydantic==2.13.1
pydantic-settings==2.11.0
pydantic_core==2.46.1
pydeck==0.9.1
pydot==4.0.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyWavelets==1.8.0
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.2.1
referencing==0.37.0
regex==2026.4.4
requests==2.32.4
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1743371105/work/dist
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==3.3.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.7
sparse==0.17.0
starlette==0.48.0
streamlit==1.56.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.20.3
toml==0.10.2
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.5.5
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traits==7.0.2
transformers==4.46.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.37.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
